[131-multilanguage-support]: Small i18n fixes and refactoring
