+++ app/components/rounded-form/RoundedForm.js
@@ -4,7 +4,7 @@ import { connect } from 'react-redux'
 import { bindActionCreators } from 'redux';
 import * as Joi from 'joi'
 
-import { i18n } from '~/i18n/i18next.config'
+import { i18n } from '~/i18next.config'
 import { RoundedFormState, RoundedFormActions } from '~/state/reducers/rounded-form/rounded-form.reducer'
 
 const inputChildComponentNames = ['RoundedInput', 'RoundedTextArea']
