+++ app/containers/Root.js
@@ -5,7 +5,7 @@ import { I18nextProvider } from 'react-i18next'
 import { ConnectedRouter } from 'react-router-redux'
 import ReduxToastr from 'react-redux-toastr'
 
-import { i18n } from '~/i18n/i18next.config'
+import { i18n } from '~/i18next.config'
 import App from './App'
 
 type Props = {
