+++ app/menu.js
@@ -1,6 +1,6 @@
 // @flow
 import { app, Menu, shell, BrowserWindow } from 'electron'
-import { i18n } from './i18n/i18next.config'
+import { i18n } from './i18next.config'
 
 const t = i18n.getFixedT(null, 'menu')
