+++ app/state/reducers/get-started/get-started.reducer.js
@@ -1,6 +1,6 @@
 // @flow
 import { createActions, handleActions } from 'redux-actions'
-import { i18n } from '~/i18n/i18next.config'
+import { i18n } from '~/i18next.config'
 
 import { preloadedState } from '../preloaded.state'
 import Wallet from '~/service/bip39-service'
